Move block device initialization code from ota.cpp to default_bd.cpp
