+++ app/ota/ota.h
@@ -72,8 +72,4 @@ struct BlockTableData {
 #define NO_OTA_FILE          (-3)
 #define INIT_FAILED          (-4)
 
-#define FILEBD_READ_SIZE      0x1
-#define FILEBD_WRITE_SIZE     0x1
-#define FILEBD_ERASE_SIZE     0x1000
-
 #endif //__OTA_H
